+++ sumo/src/microsim/traffic_lights/MSTrafficLightLogic.cpp
@@ -288,7 +288,7 @@ MSTrafficLightLogic::setCurrentDurationIncrement(SUMOTime delay) {
 void MSTrafficLightLogic::initMesoTLSPenalties() {
     // set mesoscopic time penalties
     const Phases& phases = getPhases();
-    const int numLinks = phases.front()->getState().size();
+    const int numLinks = (int)phases.front()->getState().size();
     assert(myLinks.size() >= numLinks);
     SUMOTime duration = 0;
     std::vector<SUMOReal> redDuration(numLinks, 0);
